Part 1 is done
