+++ DmitryBachurin/src/components/MessageField/Message/Message.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Message = ({name, content}) => {
+    return <li><strong>{name}:</strong> {content}</li>
+};
+
+export { Message };
